structuring project done
