+++ run.py
@@ -1,0 +1,4 @@
+from grabpage import grabUserPages
+DEBUG = False
+if __name__=="__main__":
+    grabUserPages('mineo')
